+++ python/ray/air/_internal/session.py
@@ -74,7 +74,7 @@ class Session(abc.ABC):
 
 def _get_session() -> Optional[Session]:
     from ray.train._internal.session import _session_v2 as train_session
-    from ray.tune.session import _session_v2 as tune_session
+    from ray.tune.trainable.session import _session_v2 as tune_session
 
     if train_session and tune_session:
         logger.warning(
