+++ python/ray/tune/__init__.py
@@ -8,7 +8,7 @@ from ray.tune.registry import register_env, register_trainable
 from ray.tune.trainable import Trainable
 from ray.tune.callback import Callback
 from ray.tune.suggest import grid_search
-from ray.tune.session import (
+from ray.tune.trainable.session import (
     report,
     get_trial_dir,
     get_trial_name,
@@ -39,7 +39,7 @@ from ray.tune.sample import (
 from ray.tune.suggest import create_searcher
 from ray.tune.schedulers import create_scheduler
 from ray.tune.execution.placement_groups import PlacementGroupFactory
-from ray.tune.utils.trainable import with_parameters
+from ray.tune.trainable.util import with_parameters
 
 from ray._private.usage import usage_lib
