+++ python/ray/tune/experiment/trial.py
@@ -38,7 +38,7 @@ from ray.tune.execution.placement_groups import (
     resource_dict_to_pg_factory,
 )
 from ray.tune.utils.serialization import TuneFunctionEncoder
-from ray.tune.utils.trainable import TrainableUtil
+from ray.tune.trainable.util import TrainableUtil
 from ray.tune.utils import date_str, flatten_dict
 from ray.util.annotations import DeveloperAPI
 from ray.util.debug import log_once
