+++ python/ray/tune/tests/test_function_api.py
@@ -11,8 +11,8 @@ from ray.rllib import _register_all
 from ray import tune
 from ray.tune.logger import NoopLogger
 from ray.tune.execution.placement_groups import PlacementGroupFactory
-from ray.tune.utils.trainable import TrainableUtil
-from ray.tune.function_runner import with_parameters, wrap_function, FuncCheckpointUtil
+from ray.tune.trainable.util import TrainableUtil
+from ray.tune.trainable import with_parameters, wrap_function, FuncCheckpointUtil
 from ray.tune.result import DEFAULT_METRIC, TRAINING_ITERATION
 from ray.tune.schedulers import ResourceChangingScheduler
