+++ python/ray/tune/tests/test_trainable_util.py
@@ -12,7 +12,7 @@ import ray._private.utils
 import ray.cloudpickle as cloudpickle
 from ray.tune.utils.util import wait_for_gpu
 from ray.tune.utils.util import flatten_dict, unflatten_dict, unflatten_list_dict
-from ray.tune.utils.trainable import TrainableUtil
+from ray.tune.trainable.util import TrainableUtil
 
 
 @pytest.mark.parametrize(
