+++ python/ray/tune/trainable/__init__.py
@@ -1,0 +1,19 @@
+from ray.tune.trainable.trainable import Trainable
+from ray.tune.trainable.util import TrainableUtil, with_parameters
+from ray.tune.trainable.session import Session
+from ray.tune.trainable.function_trainable import (
+    FunctionTrainable,
+    FuncCheckpointUtil,
+    wrap_function,
+)
+
+
+__all__ = [
+    "Trainable",
+    "TrainableUtil",
+    "Session",
+    "FunctionTrainable",
+    "FuncCheckpointUtil",
+    "with_parameters",
+    "wrap_function",
+]
